Allow ConnectJavaDebuggerTask to be used from any ProgramRunner
While connecting the Java debugger, we need to reuse the existing
descriptor. The descriptor was being obtained by assuming that the
program runner was AndroidProgramRunner which had a method to obtain the
descriptor. This CL obtains the descriptor via the env, thereby allowing
the ConnectJavaDebuggerTask to be reused regardless of the program
runner.

Change-Id: Ia99f53f3adb1e97ce2dfc11d006952ca6dd9f10c
(cherry picked from commit a21e0e78d9c19d037c7f58517234448b3387ce22)