Add thread state charts to Studio UI
Add a list of state charts to CPU Profiler UI level 2. These charts
correspond to the threads that are or were alive at some point since the
profiler started.

As we were relying on the assumption that each SeriesDataType would be
mapped to one data adapter, the SeriesDataStore had to be changed in
order to support 1:n (instead of 1:1) relationships with SeriesDataType.

After this change, one extra parameter (Object) can be set when
registering an adapter to the data store. Such parameter can be used
later to call the correspondent methods (e.g. getSeriesData, getDataAt),
which will then return the correct adapter.

Change-Id: I028cfe31c513d1d3e3bdb9f9345f29fe92653069