Another hack to stop the setting the atom list width to 0.
Normally the grid is restored like this:
1 - GridImpl.addNotify restores the value to 0 (as the width is 0)
2 - RunnerContentUi.MyComponent.addNotify restores the correct value
3 - RunnerContentUi.MyComponent.removeNotify saves the correct value

but sometimes the 2nd event can happen AFTER the 3rd, this causes
the 0 value to be saved, and the next time restored as the correct value.

This fix means that even if the evnts happen in the wrong order,
we never restore or save the value 0.

bug: http://b/29631719

Change-Id: I5f8de08942f0e00720d5170f1fa8a833d9c7eca9