Update default value of constraint layout auto-connect
Both the get and set methods need to be told the default
value (they use this to avoid writing values into the persistent
file that are identical to the default. This means that if
we don't update the default value provided to the setter, it
will omit it in the wrong case, which will then mean the
opposite thing to the getter on next startup.

Change-Id: Id9e40fa9bb9088712f4cbf7d87b856b8db1c87d3