Fix crash in ResourceFolderType#getTypeByName
We have the following frequent crash report:
AssertionEr @ ResourceFolderType:55 < ResourceManager:83 < :78 < :271 < :350 <
ResourceReferenceConverter:303 < :185 < GenericDomValueReference:203 <>

The root cause appears to be that somebody passes a resource type name with
a dash to it to the ResourceFolderType lookup method. This is probably
ResourceType.DECLARE_STYLEABLE, which should never have been passed to this
method. This is because the old ResourceManager just uses strings for
resource type and resource folder type parameters, and mixes & matches them
accidentally. The proper fix is to switch to the enums for this; but for 2.2,
just spot fix this.

Change-Id: Id80da03403df9097a0837d9d5cbfd714bd5ac4f6