Remove LOCAL_RES_ONLY and LOCAL_JAVA_ONLY flags
This is a temporary fix for beta1. We might remove this option entirely,
since the performance benefits of this are not worth the confusion it
causes inside Gradle.

The file change listener itself may not be needed as well - but keeping
it around in case we need it for flight recorder etc.

Fix https://code.google.com/p/android/issues/detail?id=213205 and for
other bugs where Gradle is confused if LOCAL_JAVA_ONLY is passed in but
manifest is changed via build.gradle

Change-Id: Id30aaad4ebf4d3f72b243bd09bb7c8ba38e8d280