Clamp the rendering range to not show all the way to current time
Currently, the profiler UI pops at the far right end, because we are
rendering part of the time range we haven't fetched data for yet;
therefore, when we do get it, it appears to pop into existance.
By hiding that part of the time, we give the illusion of data being
available instantly.

Change-Id: I948cc9c7121465a5fe6983cde07f12e150bd5c84