Fixing a NPE when an activity completed event is raised.
Depnding on what information the perfd has collected, mainly depending on
when it was started. It can get into a state where it only has an activity
completed event and no activity started event. The UI assumes that it wlil always
recieve a started event. This change forces us to ignore the completed event.

Change-Id: I49c52d6b95220c5068f969b203f1f218481213e8