Improve the validation for user selected launch activities
When users select to launch a specific activity (rather than just
launching the default activity), we allowed them to pick any activity
from their project. The only validation performed was to check whether
that activity is registered in the manifest.

This CL adds an additional check to also confirm that the activity
has the launcher intent specified. This is implemented as a validation
check, so the user experience is that we allow them to pick any
activity, but then show the warning that it doesn't have the right
registration in the manifest.

Fixes https://code.google.com/p/android/issues/detail?id=211354. The bug
mentions checking for exported flag, but all activities that have the
launcher intent are exported by default.

Change-Id: I238866d602744151153e2161c1f331eefb7fec41