214825: Android studio hangs for a while with code complete
Bringing up the documentation for a class (such as TextView) would
take "forever".

The root cause for this is that the reference documentation format
appears to have changed significantly recently. This has not yet been
published to the SDK documentation component, so if you have
documentation installed locally, you don't observe this problem.

In the IDE we have a custom javadoc external filter which goes and
attempts to extract out just the relevant signature and class
description text, and skip all the other stuff (such as the massive
constructor, field and method tables). In addition, the javadocs now
also contain a bunch of other text (such as the option to switch to
other languages, and these were included using glyphs in the various
languages, which meant the HTML renderer would end up having to look
up the font glyphs for these as well).

The fix is straightforward: the documentation filter now looks for
tokens both in the old format and in the new format.

This CL also adds a unit test for the external filter.

Note that documentation for some of these classes can't be instant --
the TextView.html file itself is over 1 MB large!

Change-Id: I3183c730106013aee5263bc5cbea1667315989f6