Fix Issue 215756: Bug with pan and zoom for XML preview in AS
The bug occured because the widget was waiting for the componentListener attached to the design surface to be triggered, but was never actually called.
Replaced it with an AncestorListener attached to the DesignSurface that is actually called when the DesignSurface is hidden.

Also made sure that the content of the widget is updated when switching to another editor using the same DesignSurface

https://code.google.com/p/android/issues/detail?id=215756

Change-Id: Idde33f9099887d452ac795fb9688285e1d98c6e7