210458: Update the device graphics for Nexus 7 v2 and Nexus 10
These two devices had much bigger screen resolutions than the earlier
phone assets, so instead of a full-resolution image, the Nexus 10
had a 1/2 scaled image, and the Nexus 7 had a 2/3 scalde image. The
layout editor dealt with this without a problem (because it scales
the assets to match the device metadata provided by sdklib) but the
emulator skins did not.  To fix this, the images have been scaled up
(200% and 150% respectively) and the metadata coordinates updated to
match.

Change-Id: I50ac1cae0857e4d80a408a7c870d4d4e7d990a3d