Add Checked to Switch inspector.
Bug: http://b.android.com/214081

J requested that the Switch component inspector added both
android:checked and tools:checked per bug 214081.

In order to support this in ViewInspectorProvider we can no
longer hold the editors in a Map from property name.
Changed the Map to a List and introduced method for getting
all editors of an ComponentInspector.

The method activatePreferredEditor will use the ViewManager
to decide the preferred property to give focus.

Ideally the method: updateProperties should not recreate
the editors but just update the property values.
This will be addressed in another CL.

Change-Id: I0f16d994b55155317202cc2cf9335b8b7ba11de6