Fix IDs refresh when adding new elements
This fixes the case where you have a widget with an ID, remove it from
the XML and then undo the operation. Before this change, the undo
operation will not refresh the re-added ID and will leave the resource
repositories in an inconsistent state (since it will be missing from the
caches).

Change-Id: I621bb128c5e6712ba8abda056c38c4bc7a56fbda
(cherry picked from commit 5dc1d4f1a49b6e6845a08606158bd99c0185edcf)