Core Shaders Tab Implementation.
Includes making calls to fetch the list of shaders and programs
and from there making calls to fetch shaders associated with programs.
Displayed fetched shader source in the mySourcePanel component.
Created TextCellController, which serves as a general-purpose,
AWT-backed list of Strings that can be retrieved from an Rpc.listen call.

Change-Id: I5353b6679fd487f64df93ea2a66df93405541fb7