Activate editor through a DesignSurfaceListener.
bug: http://b.android.com/216935

A double click on a TextView in the designer should cause the
text attribute to be selected in the inspector.
Prior to this CL this only worked if the property tool window
was detached.
With this CL is also works if the property tool window is
attached, and will switch to the inspector if the property
sheet is currently selected.

Change-Id: Ib4b184f736100d459f92eab1a7a85667c55a4716
(cherry picked from commit 85946d83a7d6ef32e90b32ef0a21b6f67b711c2e)