Allow layout inspector to be launched as an action
This CL enables layout inspector to be launched as an action via a
keyboard shortcut (Cmd+Shift+A, search Layout Inspector). It is also
added to the menu in Tools | Android.

When invoked in this way, we show a process chooser dialog, and then
connect to the selected process. This CL refactors the existing process
chooser dialog so that it can be showed in different contexts (attach
debugger, or attach layout inspector).

There are also a few other cleanups to the process chooser dialog - the
rendering of the device and client have been improved, and the clients
are displayed sorted by their package name.

Fixes:
https://code.google.com/p/android/issues/detail?id=203230
https://code.google.com/p/android/issues/detail?id=213356

Change-Id: Idf57632457ea9c63cedaae592ab4892b59ef7f4a