layout inspector: make sure the two panels on the side have width > 0
This is a partial fix for Issue 210440: sometimes the tree and
properties panel are invisible. It turns out that once in a while, they
somehow get their width set to zero.

A proper fix will be in 2.3 where both of these will become
LightToolWindows similar to the layout editor.

Partial fix https://code.google.com/p/android/issues/detail?id=210440

Change-Id: I9efd7ddfda8a4aba50a8e7934c4aee61205da841