EmulatorConnectionListener: Tweak to work when adb integration is off
When deciding whether an emulator has fully booted up, we use two
heuristics: the dev.bootcomplete property is not null, and that the
emulator has debuggable apps that are running.

However, the second condition only makes sense when adb integration is
turned on. If it was turned off, then we'd not know about any debuggable
clients, and hence wait forever thinking that the emulator hasn't
finished booting.

This CL changes the implementation to perform the second check only if
adb integration is enabled. This is ok because the second heuristic is
not strictly necessary with the addition of dev.bootcomplete flag.

Fixes https://code.google.com/p/android/issues/detail?id=209095

Change-Id: Ica207abea261943cffe3fbe22655252db1b446ba
(cherry picked from commit f882e19dd195ce9f49cc1077507cdc0cf46ecae3)