Give AVDs cameras by default
When a new Android Virtual Device is created, the user can select
if the device has a front and/or back camera. The choices are
None, Emulated, and <Connected physical device>.

The default was None; it is now Emulated.

BUG: b.android.com/210175

Change-Id: Icfaccb449e91c02b6a4b15f54ab9ff8b5e3b5589
(cherry picked from commit 4597d7e14a7888295bb459dc26b72ad409030e1f)