Provide more information and stop launch on IR errors
An assertion that seems to indicate that we are seeing hotswap artifacts
for a cold swap build seems to be getting triggered every now and then.
Rather than terminate in an assertion error, we now explain what is
going on, log an error and terminate the launch when this happens. The
error log has additional details which will allow us to confirm the
issue.

https://code.google.com/p/android/issues/detail?id=217934

Change-Id: Ibb18797120f23e0304531aa06c28bdc26570cb6a