Force kill app when assuming that the app is not running
If the IDE encounters an error while trying to detect that the app is
running, it assumes that the app is not running. But if this happened
due to an error (such as port forwarding error due to address being in
use), then this assumption is wrong (we'll generate freeze swap patches
and start the app, but the app won't actually be restarted because it
was always running).

This CL avoids this issue by making sure to kill the app when we do run
into such cases (which should be pretty rare).

Fixes https://code.google.com/p/android/issues/detail?id=218593

Change-Id: I0c12f49c33271e8c9bbc8cfa36d774b772c36bc7
(cherry picked from commit 35bfc2fafa9d0e3576bbac583bb09275be136be3)