Don't update Gradle to 2.14.1 if using stable gradle plugin
Continuing the fix in an earlier CL (fcfb2ddcabb8), this CL stops the
IDE from offering even an update of Gradle itself to 2.14.1. With this
CL, users on 2.1.x stable plugin of Gradle will not be offered an update
on either the plugin or gradle for security reasons.

There is still a separate flow that always suggests an update to the
latest version of the gradle plugin (2.2.0-alpha7), which still remains
as it should.

Continuing fix for
https://code.google.com/p/android/issues/detail?id=218614

Change-Id: Ib8d7e78fe6a448a8531c7e3f0c139ca1bb601523