Allow launching activities with any intent filter
A while ago we tightened the requirement for what activities can be
launched (4a19716434). This is another follow up CL to relax that
restriction. In this CL, we relax the requirement that there be an
explicit LAUNCHER intent: an activity is exported as long as it has
*any* intent filter, and any exported activity can be launched.

Fixes https://code.google.com/p/android/issues/detail?id=218981

Change-Id: Ic067fe39143155712d5068ce42cb734161cd8538