Adding listener to update the VFS when [un]installs finish
In Studio, we create Sdk objects that point to each platform that
has ever been installed or referenced, regardless of wether it's
currently installed or not. Those Sdks include VirtualFile references
to the actual SDK component contents, if it exists.
When we change (and especially delete) installed SDK components, we
need to go and refresh the VirtualFiles referenced in the Sdk objects.
We need to ensure this happens after the VFS entries for the modified
components is complete, which includes both the refresh of the files
themselves and the asynchronous refresh of jar file contents that is
kicked off by the initial refresh, so we are able to tell what has
actually changed.
In the end, the state of an Sdk object after e.g. deleting the
underlying SDK component should be the same as if we started up
without that coponent installed at all.

Fixes b.android.com/215823

Change-Id: I295838823202c7dd5a6e0e5d0d401a72455e5291