Refactor DragHandler.commit to also add the components.
The code to insert the NlComponent into the layout after a
drag and drop operation was done in the DragDropInteraction
class. But it seems better to have the commt() method in
the DragHandler do this because the ViewHandler that creates
the DragHandler can then customize how the components being
dragged will be inserted.

Change-Id: I3613fc54ac427f141abf24f7e0a09a4f96f99edf