[New Sync] Simplified API for setting up content roots.
Refactored code by adding JavaContentEntries and
AndroidContentEntries. These classes do the same as
ContentRootModuleSetupStep (Java and Android
respectively) with the difference that they
have the list of ContentEntry, the ModifiableRootModel
and the respective build models as part of their state,
instead of passing them around as method parameters,
resulting in code that is easier to read.

Change-Id: I97a6749a9e28dc13b097caf995a0c75715b63420