Fix text state chart image diff test
Use a TrueType Font to avoid large render differences across different OS.
The goal of the test is to check that text is being rendered (instead of
color bars). Therefore, it's OK if the font rendered in the test is
different than the one that is actually going to be rendered in the
application.

Bug: http://b.android.com/219264
Change-Id: I7edd64230580496a4441c941b49729ba697eb95f