Make sure we dont update properties from 2 different threads.
Bug: http://b.android.com/219552

The method updateDefaultProperties was called from a non UI thread
and it was iterating over a list that was initialized on the UI
thread i.e. it was possible that the list was modified in the
middle of the other thread was iterating over the contents.

Change-Id: Ib14646673e0a48f7fe3d5eb6fa3bcb7f5183a296
(cherry picked from commit 7f1d1d354d003ca6debdf1828add806575c4ec77)