Exports asm from android build
In order to work around a Bazel limitation where the exports are
inserted at the beginning of the class path and given that android
depends on asm4 and asm5 at the same time. We need to make sure
the classpath is in the right order. This CL achieves that
by exporting the asm4 package in the right order.

This will be removed once the Bazel export logic is sorted out.

Change-Id: I66730b83b2b5c4fe84a44990e63a354cc88e7d88