Replaced utility methods in GradleUtil with real objects.
GradleUtil has become a disorganized repository of utility
methods that are related to Gradle. As time passes the
class is growing bigger and it is hard to tell what methods
this class contains.

This CL moves a few of these methods to objects with a
single responsibility:

- AndroidPluginGeneration: indicates whether the Gradle
  plugin used in a project is the stable or experimental
  plugin.
  It also indicates which is the "recommended" version of
  a plugin generation.
- AndroidPluginInfo: information about the Gradle
  plugin used in a project (e.g. version, "app" module,
  generation.)
- BuildFileProcessor: processes build.gradle recursively.
- GradleCache: related to the local Gradle cache. It
  can find artifacts in the cache.

This CL is pretty big because all this new classes are
related.

This CL also introduces integration tests using the
experimental plugin.

Change-Id: Ia5fb80e83a7399a318c79220b297c3ca25395be4