Updating the constraintlayout lint check
Changed the constraintlayout fix to just update the version to the
latest (local or remote). Gradle will then prompt for download if
needed.
There was also a problem in GradleDependencyManager where we wouldn't
consider dependencies for which we had no artifacts installed as unmet,
and thus we wouldn't try to download them. Fixing that.

Part of the fix for b.android.com/219556

Change-Id: I0900b7fa5066554387e33c195318a8aa6c2aa492