pkginfo: cache results, support receiving icons in second pass
Adds a package info cache keyed by device serial number.
When the dialog pops up, immediately populates the list with
information retrieved from the cache. A spinner is shown in
the bottom-left of the dialog, and the package information is
shown in grey to make it obvious that processing is still under way.
If the user grabs their package from the list, they can start
the trace without waiting for the tree to finish populating.

Adds support for receiving the information in two passes, which
allows us to immediately show the textual information, and save
the few seconds it takes to grab the actual icons.

Bug: 29234827
Change-Id: I1f3fa9400500c2b4ae539e28a1a0e54eaf2b8819