Make PreferenceScreenDragHandler act like the category one
PreferenceCategoryDragHandler should be doing the painting when the
mouse pointer is "inside" a category and in the spacing between adjacent
preferences. DragDropInteraction selects a PreferenceScreenDragHandler
in that case because the category bounds don't contain the children and
because of the spacing.

I fixed it so it does the drag handling that
PreferenceCategoryDragHandler would have done.

Change-Id: I3f061d14db3c168a13e973f5837d9fc5a894e2c7