Changed behavior of component tree
The component tree now accepts a view to be dragued on another view
which is has exactly the tag View and the attribut mockup and is not a
descendent of the dragged view.
If the drop occurs, the target view is morphed into a ConstraintLayout
and the dragged view hierarchy is placed below the morphed component.
Only some attributes are kept when the view is morphed.

Change-Id: I1e776b6b656f9dbe107c5087e2939bb49e669318