Making dependency presence checks forgiving.
Making checks on the presence of a dependency more forgiving
by allowing checks for older versions to return true when a
newer version is found.

Justification: Appears that adding dependencies did not work. This
is a blocker for advertising the plugin.

Risk: Relatively minor as primarily a refactor but non-zero.

Bug: 31047048
Change-Id: Ida8aaedd09564bb479bc13b871902384dca0e4c8
(cherry picked from commit d0982a4570d2508b809453bb4e960324c130f2c2)