Add image diff tests related to Range
Range implements the Animatable interface and, as such, it animates over
time. Ranges are created with a min and a max and display the data added
between them. However, they should increase their max value in a visually
smooth way when a value higher than current max is inserted.

AnimatedTimeRange is a special type of range and animates over time. It
sets the max of the range for the current time in each animate call.

The tests added cover horizontal and vertical ranges, making sure the
ranges display only the data between their min and max, as well as they
increase their max once a higher value is added.

A couple of tests for AnimatedTimeRange are also included, highlighting
it animates over time.

Change-Id: I74480fcc7f73eff478a0ee05c33bc28cad55b489