Install platform corresponding to device before invoking Gradle
With commit 353f46cbc7363, Gradle will perform instrumentation for
instant run using the android.jar corresponding to the API level of the
target device. If this platform is missing in the SDK, the build will
fail. To avoid this scenario, the IDE will have to detect whether the
platform is installed, and if it is not installed, prompt the user with
the following options:
 - either download & install the platform and use instant run.
 - continue without instant run.

There is still one remaining issue though: Gradle seems to automatically
do a sync everytime a platform is installed. This comes from the VFS
listener installed in ac951889030012cb. This should be removed or
updated appropriately. This is tracked by http://b.android.com/220945
(Note: the above issue is only present in master, not 2.2)

IDE side fix for http://b.android.com/220852

Change-Id: I90e53fe6cb96a84ef9f8519311af796d08af5106