Use GradleSyncListener instead of project save listener in palette.
Use a gradle sync listener instead of a project save listener in
a Nele palette. We want to update the missing dependencies from
the possible palette components.
We used to to update those on project save notifications.
However Gradle sync events are better.

Change-Id: I93715973fac828251323a3ae6717f27ee84f60f5
Also: make sure to repaint the palette after such a notification.