Avoid leaking DynamicResourceValueRepository
Currently, when a new DynamicResourceValueRepository is created, it is
subscribed to two listeners to detect Gradle syncs. When refreshing a
layout (pressing 'r'), all the repositories are discarded but they
remain in memory since they are referenced as listeners.
This CL removes the references from the listeners and also registers
them with the Disposer.

Bug: http://b.android.com/221302
Change-Id: I9e071a6fec580d248689a8b83d6c845a2714e77c