Add support annotations external zip file during Gradle sync
The support annotations library is just a .jar file, not an AAR
(deliberately, such that it can be used outside of Android library
projects too). But that means it can't provide typedef annotations
in the AAR the way we do it for android libraries. Therefore, in
addition to shipping -sources.jar and -javadoc.jar, the support
annotations library includes a -annotations.zip file, and during
sync we need to register this root as an external annotations root.

Change-Id: I73f579f1f3302506e1143c69cea69551e1c320d5