Fix 2 layout editor bugs
Bugs: http://b.android.com/221531  http://b.android.com/221565

1) Styles and TextAttributes
When editing styles (and TextAttributes) the display text is
shortened from the actual attribute value.
When a user types a value (or the control loses focus) we took
the shortened form and updated the XML with this value.
Now we attempt to lookup the style value from the styles we are
already presenting as dropdown choices.
If we do not find it assume it is a project style and add:
"@style/".

2) Focus problems in property table
We recently added code to stop editing when focus is lost.
The property table is using a JBTable which is attempting to
override the focus management to the editor. Our editor is a
JPanel with several components on it. The JBTable is causing
our combobox to lose focus as soon as it receives it.
This CL eliminates that by overriding the method:
JBTable.surrendersFocusOnKeyStroke

Change-Id: Ia1baeb7c1fe015250ad1469c6adf01fc5340ff16