Introduce settings to choose between two types of data binding classes
We generate in memory classes for data binding layouts which speeds
up development a lot. Unfortunately, they hide the generated classes
by the compiler so if you have something going wrong in your data
binding expressions, it is not possible to debug the generated code.

This CL adds an option to choose between these two options so that
developers can debug the generated code when needed while enjoying
the convenience and speed of the in memory codes.

I've also tried toggling this value based on Debugger state but
it confuses AS a lot and cannot provide a predictable user experience.

Bug: b.android.com/208975
Change-Id: Ieee2c352094aa91f53a3576862f761b0b9a3a1db