Remove duplicate layout set code for myContents.
Cherry-pick justification: Swing layout fix that corrects Firebase Assist tutorial text that was being clipped.
Risk: Very low as the only change is the removal of a second setLayout call.

VerticalFlowLayout seems to be the correct layout to use for myContents,
but it was being changed to BoxLayout later for some reason.

This seems to fix most of the width clipping issues on the elements, but
there does seem to still be a minimum width being enforced on step elements.
It seems to either be the minimum width of the tutorial title or
the Launch in Browser link/editor.

Bug: 31264330
Change-Id: I54a8f146d287c10bb13a9748b4bd19c36014e7ba
(cherry picked from commit f5aaf514dd5f79f05c0d03c9394124701441df39)