Fix copy, paste, undo & redo in NlPreviewForm
We delegate the handling of those actions now to the DesignSurface
instead of the NlEditorPanel so they can be reused across the preview
and the design views.
The undo & redo actions are actually handled by a FileEditor so in the
case of the preview, we delegate them to the selected TextFileEditor
that is displaying the XML.

This CL also updates the handling of paste so, when there are no
components selected, we just use the root of the currently displayed
layout. Before, we wouldn't paste the content unless something was
selected.

Change-Id: I8ca84d97ce0807fade9d788efe56bd03ac9f5fa6