Improvements to VersionCompatibilityChecker.
(This class is the one that verifies version
compatibility between IDE components. The
compatibility checks are declared in metadata,
making it easy to update this checks for older
Studio releases.)

This CL splits VersionCompatibilityChecker into
smaller classes with a single responsibility.
The code is now simpler and easier to maintain
and test.

This CL fixes a couple of minor bugs and adds
tests. This CL also converted related UI tests
to headless integration tests.

Change-Id: Id45133c88dbcebdbb4c6e92055fb4c5ad4e2b31c