Introduce legacy version of Android test base classes
We have nearly 200 tests depending on these base classes, but we want
to fix them up to rely on a prebuilt test SDK, not an arbitrary SDK
pointed at by an ANDROID_HOME env var.

By forking legacy versions of these base classes and updating all
tests to rely on them, we can start to tinker with the new versions
of these base classes and convert tests over to the new versions
incrementally.

Change-Id: I9d99e0faee1cbb445100510935d45dad0879be4e