Fixed transitive dependency resolution.
Transitive dependency resolution was broken because the
IDE expected the Android plugin to return, during sync,
the full dependency graph. This feature has been disabled
already in the plugin, but the IDE was not updated.

This CL removes the code that handled the full dependency
graph and restores the previous behavior where only
first-level dependencies are returned, and each module has
to "export" their dependencies to make them visible to
dependent modules.

Fixes:
https://code.google.com/p/android/issues/detail?id=222926
https://code.google.com/p/android/issues/detail?id=222924

Partially fixes:
https://code.google.com/p/android/issues/detail?id=221883

Change-Id: I38115d6e3a96e45b34d5fb4500db00e3ed44c17f