220197: Add a search box when choosing a vector material design icon
First, add a search text field above the icon table. The search
filters the list for any case insensitive substring match of the
filenames. Also, add a keylistener to the table such that if you start
typing in the table, it transfers the character and the focus into the
search field to simulate filter-search behavior.

Second, tweak the cell renderer such that it also displays the names
of the icons - this is the subject of issue 186432. These are based on
the filenames, by stripping off the ic_ prefix, dropping the
_black_24dp.xml suffix, and replacing underscores with spaces. This
seems to be similar to what is shown in
https://design.google.com/icons/

Third, improve the handling of dark theme: When using Darcula, apply a
color inversion filter to the icons such that they show up light gray
on dark instead of black on dark gray; this also lets us turn the
table background back to dark. (This is issue 199881.)

Fourth, the Material Design icons have been updated to the latest
version (3.0.1), which includes changing the license to Apache 2.0, so
update the hyperlink and text for the license.

Also fix misc warnings and other cleanup.

Change-Id: I20d5d3a79d10767cd797c12ae0544b86759e95fd