Allows dev jdk to be specified as a System property.
Instead of just looking at "../../prebuilts/studio/jdk",
we now allow this directory to be set via a system property.

This is needed because when this directory is a symlink,
gradle fails at recognizing that we use the same jdk in
two invocations as its resolver is bogus and fails to
do it correctly. See:

https://discuss.gradle.org/t/gradle-daemon-different-context/2146/3

Change-Id: I50b6a036cb556307eace2b175d1aac209c21f3a5