Avoid XML update when updating the properties panel
When moving around a widget requires updating the properties panel, it
will usually trigger an XML save even if the interaction is not
finished. That will cause a full refresh and a layoutlib call, making
the UI stutter when dragging.
Since the properties panel update is being triggered because of a widget
update, there is no need to feed that change back to the widget (since
it's the original source). That avoids the XML save.

Also removing some unused code in the WidgetConstraintPanel

Change-Id: Ie8ac4780f1c7e1c41fb37adba1cae1c4001872c3