Update comment for TestProjectPathsGeneratorTest exclusion
Most tests in our blacklist are for testing Android Studio; however,
we learned that TestProjectPathsGeneratorTest is a test-only project
used for generating metadata for other unit tests.

Ultimately, we should rewrite the class to be bazel friendly, but it's
less urgent for now, so we indicate that by moving it into a different
area of "todo later" blacklisted tests.

Change-Id: I5f722d02f56aa28667abbb885104e6f0ab2ad30f