Add logcat content to flight recorder
This CL adds the output of the instant run runtime library to the flight
recorder logs. This works as follows:

 - We make sure that the instant run logs are enabled at verbose level
 - all logs from the runtime library are saved in an in memory ring
   buffer
 - whenever the submit feedback option is selected, the contents of the
   buffer are saved to a file and uploaded along with the other logs

Currently, we don't stop listening to logcat when the app dies or if the
device is disconnected.

(cherry picked from 6ea3543dd494ed23a7cd94ded2ae97d9e5df64ea)
Change-Id: I42b28643cfb9e825d2131e56f27e3eafe6f7ea25