Add Pixel & Pixel XL device frame art
The pixel assets are in WEBP format since they're really large
(15MB for the PNG assets, less than 10% of that for WEBP). These
assets get rewritten to PNG on demand when the skin files are
created for the emulator (which will start to support webp in
version 25.3).

Also updates tests for the new Pixel devices.

Change-Id: I32d47dd9635b16434fd0c90c98441f8385f3fe2f