Fix UI of FlatButton and FlatComboButton to work on all L&F
The way the border was set on those buttons was not working well
with the Windows L&F, as the Windos L&F sets the border to its
default value no matter if a custom one already exists.
In addition, the font of FlatButton is updated to match the one of
FlatComboButton.

Change-Id: I10ae268625e78daa1eee095a8e44fce818f6cff8