Fixed bug where source folders for native projects where not set.
The problem was that ContentRootModuleSetupStep was
looking for NativeAndroidGradleModel directly in the
facet, but the changes in the facet have not been
committed yet and the facet has not been set yet.
Because there is no "native" model, sync removes all
cpp source folders.

The solution is to look for the facet first in
IdeModifiableModelsProvider (where changes that have
not been committed yet are stored) and then look
for NativeAndroidGradleModel in the facet.

Fixes:
https://code.google.com/p/android/issues/detail?id=224587

Change-Id: Ia56c7d0e479aed19ed3af0f786f0523957c044c1