Ensures the palette is only listening to one configuration
It used to be that the palette was associated to the current
configuration of the design surface. But if setConfiguration
was called on the design surface, it would change the configuration
of the surface without updating the palette, and the old
configuration would forever have the palette as a listener.
We now only use the configuration field from the palette, making
sure that the palette is associated with at most one configuration
at a time.

Change-Id: Ieae1134f9a045f97ce00c45c27819ee628215a15