Wire error reporter to use Google Crash instead of Feedback
The android plugin registers an error reporter. This was earlier wired
up to report to Google Feedback. This CL changes the back end to be
Google Crash instead.

In addition, when users have allowed sending of statistics, the platform
will now push crash, exception and anr reports via the same error
reporter extension to the android plugin. These are also wired up to
upload to Google crash. (We use this mechanism of platform pushing to
android plugin via the error reporter extension so that the Google Crash
code can live in a single place - in the Android plugin).

Change-Id: I4907881200aeeb0091b1bcf9f28fca959b869151